Move CheckstyleActions interface to csapi package
