+++ src/csaccess/java/org/infernus/idea/checkstyle/service/CheckstyleActionsImpl.java
@@ -9,10 +9,10 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiFile;
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
-import org.infernus.idea.checkstyle.CheckstyleActions;
 import org.infernus.idea.checkstyle.checker.CheckStyleChecker;
 import org.infernus.idea.checkstyle.checker.Problem;
 import org.infernus.idea.checkstyle.checker.ScannableFile;
+import org.infernus.idea.checkstyle.csapi.CheckstyleActions;
 import org.infernus.idea.checkstyle.csapi.CheckstyleInternalObject;
 import org.infernus.idea.checkstyle.csapi.ConfigVisitor;
 import org.infernus.idea.checkstyle.csapi.TabWidthAndBaseDirProvider;
