+++ src/main/java/org/infernus/idea/checkstyle/csapi/ConfigVisitor.java
@@ -2,6 +2,7 @@ package org.infernus.idea.checkstyle.csapi;
 
 import org.jetbrains.annotations.NotNull;
 
+
 /**
  * Visitor called by the PeruseConfiguration command.
  */
